Optimize searching
Instead of deleting all the widgets, then creating and displaying ones that match search string, now search function moves the widget that don't match to the bottom of the list, disables them, and changes their opacity to 0, so they are invisible. This approach is much faster, as widgets are always stored in memory, and only their opacity and position changes.

Write to file after every operation
When the app would crash all changes would lost as previously changes were only made when user exited the app properly